LiveEmotion/UnifiedEmotion/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
torchaudio
librosa
soundfile
python-multipart
numpy
opencv-python
Pillow
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
